requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
PyJWT==2.8.0
Werkzeug==2.3.7